requirements.txt
discord.py[voice]

fastapi==0.111.0
pydantic>=2.7.0,<3.0.0